requirements.txt
pandas==1.1.5
requests==2.27.1
openpyxl==3.0.10
python-dotenv==0.20.0
xlrd==2.0.1
xlwt==1.3.0
streamlit==1.10.0 